identifier function and related code to deprecated
