+++ src/deprecated.jl
@@ -1,3 +1,4 @@
+import Base: isidentifier, is_id_start_char, is_id_char
 import Base: @deprecate
 
 @deprecate by(d::AbstractDataFrame, cols, s::Vector{Symbol}) aggregate(d, cols, map(eval, s))
@@ -765,6 +766,51 @@ function builddf(rows::Integer,
     else
         return DataFrame(columns, o.names)
     end
+end
+
+const RESERVED_WORDS = Set(["local", "global", "export", "let",
+    "for", "struct", "while", "const", "continue", "import",
+    "function", "if", "else", "try", "begin", "break", "catch",
+    "return", "using", "baremodule", "macro", "finally",
+    "module", "elseif", "end", "quote", "do"])
+
+function identifier(s::AbstractString)
+    s = Unicode.normalize(s)
+    if !isidentifier(s)
+        s = makeidentifier(s)
+    end
+    Symbol(in(s, RESERVED_WORDS) ? "_"*s : s)
+end
+
+function makeidentifier(s::AbstractString)
+    (iresult = iterate(s)) === nothing && return "x"
+
+    res = IOBuffer(zeros(UInt8, sizeof(s)+1), write=true)
+
+    (c, i) = iresult
+    under = if is_id_start_char(c)
+        write(res, c)
+        c == '_'
+    elseif is_id_char(c)
+        write(res, 'x', c)
+        false
+    else
+        write(res, '_')
+        true
+    end
+
+    while (iresult = iterate(s, i)) !== nothing
+        (c, i) = iresult
+        if c != '_' && is_id_char(c)
+            write(res, c)
+            under = false
+        elseif !under
+            write(res, '_')
+            under = true
+        end
+    end
+
+    return String(take!(res))
 end
 
 function parsenames!(names::Vector{Symbol},
